fix: rename method and delete searchModeStore[WTEL-6415](https://webitel.atlassian.net/browse/WTEL-6415)
